rag_app/requirements.txt
fastapi
uvicorn
openai
PyMuPDF
faiss-cpu
python-multipart